Develop

+++ src/main/java/dts/logic/EnhancedItemsServiceImplementation.java
@@ -3,6 +3,8 @@ package dts.logic;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -11,12 +13,16 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import boundaries.ItemBoundary;
-import constants.Constants;
 import dts.converter.ItemConverter;
 import dts.dao.IdGeneratorDao;
 import dts.dao.ItemDao;
+import dts.dao.UserDao;
 import dts.data.IdGeneratorEntity;
 import dts.data.ItemEntity;
+import dts.data.UserEntity;
+import dts.data.UserRole;
+import dts.utils.ValidationService;
+import exceptions.InvalidItemTypeException;
 import exceptions.ItemNotFoundException;
 import models.operations.CreatedBy;
 import models.operations.ItemId;
@@ -28,31 +34,46 @@ public class EnhancedItemsServiceImplementation implements EnhancedItemsService 
 	private ItemDao itemDao;
 	private ItemConverter itemConverter;
 	private IdGeneratorDao idGeneratorDao;
+	private UserDao userDao;
+	private ValidationService validationService; 
+
 
 	@Autowired
-	public EnhancedItemsServiceImplementation(ItemConverter itemConvertor, ItemDao itemDao, IdGeneratorDao idGeneratorDao) {
+	public EnhancedItemsServiceImplementation(ItemConverter itemConvertor, ItemDao itemDao,
+			IdGeneratorDao idGeneratorDao,
+			UserDao userDao,
+			ValidationService validationService) {
 		this.itemConverter = itemConvertor;
 		this.itemDao = itemDao;
 		this.idGeneratorDao = idGeneratorDao;
+		this.userDao = userDao;
+		this.validationService = validationService;
 	}
-
+		
 	@Override
 	@Transactional
 	public ItemBoundary create(String managerSpace, String managerEmail, ItemBoundary newItem) {
 		ItemEntity newItemEntity = this.itemConverter.toEntity(newItem);
-		// generate new id for item
+		
+		Optional<UserEntity> managerEntity= this.userDao.findById(new UserId(managerSpace, managerEmail).toString());
+		this.validationService.ValidateUserFound(managerEntity, managerEmail);
+		this.validationService.ValidateRole(managerEntity, UserRole.MANAGER);
+
+		if(newItemEntity.getType().isEmpty() ||
+		   newItemEntity.getType() == null)
+		{
+			throw new InvalidItemTypeException("Invalid type: " + 
+					newItemEntity.getType() + " for item: " + 
+					newItemEntity.getName());
+		}
 		IdGeneratorEntity idGeneratorEntity = new IdGeneratorEntity();
 		idGeneratorEntity = this.idGeneratorDao.save(idGeneratorEntity);
 		Long numricId = idGeneratorEntity.getId();
 		this.idGeneratorDao.deleteById(numricId);
 		String strId = "" + numricId;
-		// set the time stamp
 		newItemEntity.setCreatedTimestamp(new Date());
-		// set the item id
 		newItemEntity.setItemId(new ItemId(managerSpace, strId).toString());
-		// set the created by
 		newItemEntity.setCreatedBy(new CreatedBy(new UserId(managerSpace, managerEmail)).toString());
-		// save new item to db
 		itemDao.save(newItemEntity);
 		return this.itemConverter.toBoundary(newItemEntity);
 	}
@@ -61,11 +82,18 @@ public class EnhancedItemsServiceImplementation implements EnhancedItemsService 
 	@Transactional
 	public ItemBoundary update(String managerSpace, String managerEmail, String itemSpace, String itemId,
 			ItemBoundary update) {
-		Optional<ItemEntity> existingItem = this.itemDao.findById(itemSpace + Constants.DELIMITER + itemId);
-
-		if (!existingItem.isPresent())
+		Optional<UserEntity> managerEntity = this.userDao.findById(new UserId(managerSpace, managerEmail).toString());
+		this.validationService.ValidateUserFound(managerEntity, managerEmail);
+		this.validationService.ValidateRole(managerEntity, UserRole.MANAGER);
+		Optional<ItemEntity> existingItem = this.itemDao.findById(new ItemId(itemSpace, itemId).toString());
+		if (!existingItem.isPresent() || 
+				itemId.isEmpty() ||
+				itemId == null ||
+				itemSpace.isEmpty() ||
+				itemSpace == null)
+		{			
 			throw new ItemNotFoundException("item with id: " + itemId + "and space: " + itemSpace + " does not exist");
-
+		}
 		ItemEntity existingEntity = existingItem.get();
 		ItemEntity itemEntity = this.itemConverter.toEntity(update);
 		itemEntity.setCreatedTimestamp(existingEntity.getCreatedTimestamp());
@@ -85,7 +113,7 @@ public class EnhancedItemsServiceImplementation implements EnhancedItemsService 
 	@Override
 	@Transactional(readOnly = true)
 	public ItemBoundary getSpecificItem(String userSpace, String userEmail, String itemSpace, String itemId) {
-		Optional<ItemEntity> existingItem = this.itemDao.findById(itemSpace + Constants.DELIMITER + itemId);
+		Optional<ItemEntity> existingItem = this.itemDao.findById(new ItemId(itemSpace, itemId).toString());
 		if (!existingItem.isPresent())
 			throw new ItemNotFoundException("item with id: " + itemId + "and space: " + itemSpace + " does not exist");
 		return this.itemConverter.toBoundary(existingItem.get());
@@ -94,24 +122,61 @@ public class EnhancedItemsServiceImplementation implements EnhancedItemsService 
 	@Override
 	@Transactional
 	public void deleteAll(String adminSpace, String adminEmail) {
+		Optional<UserEntity> existingAdmin = this.userDao.findById(new UserId(adminSpace, adminEmail).toString());
+		this.validationService.ValidateUserFound(existingAdmin, adminEmail);
+
+		this.validationService.ValidateRole(existingAdmin, UserRole.ADMIN);	
 		this.itemDao.deleteAll();
 	}
 
 	@Override
-	public void bindItemToChildItem(String managerSpace, String managerEmail, String itemSpace, String itemId) {
-		//TODO
+	@Transactional
+	public void bindChild(String managerSpace, String managerEmail, String itemSpace, String itemId, ItemId item) {
+		Optional<UserEntity> manager = this.userDao.findById(new UserId(managerSpace, managerEmail).toString());
+		this.validationService.ValidateUserFound(manager, managerEmail);
+		this.validationService.ValidateRole(manager, UserRole.MANAGER);	
+		ItemEntity parentItem = this.itemDao.findById(new ItemId(itemSpace, itemId).toString())
+				.orElseThrow(() -> new ItemNotFoundException(
+						"item with id: " + itemId + "and space: " + itemSpace + " does not exist"));
+		ItemEntity toBindItem = this.itemDao.findById(item.toString()).orElseThrow(() -> new ItemNotFoundException(
+				"item with id: " + item.getId() + "and space: " + item.getSpace() + " does not exist"));
+		toBindItem.addParent(parentItem);
+		parentItem.addChild(toBindItem);
+		this.itemDao.save(parentItem);
 	}
 
 	@Override
-	public ItemBoundary[] getAllItemChildren(String userSpace, String userEmail, String itemSpace, String itemId) {
-		//TODO
-		return null;
+	@Transactional(readOnly = true)
+	public List<ItemBoundary> getAllChildren(String userSpace, String userEmail, String itemSpace, String itemId) {
+		ItemEntity parentItem = this.itemDao.findById(new ItemId(itemSpace, itemId).toString())
+				.orElseThrow(() -> new ItemNotFoundException(
+						"item with id: " + itemId + "and space: " + itemSpace + " does not exist"));
+		Set<ItemEntity> itemChildren = parentItem.getItemChildren();
+		return itemChildren.stream() // Stream<ItemEntity>
+				.map(new Function<ItemEntity, ItemBoundary>() {
+					@Override
+					public ItemBoundary apply(ItemEntity entity) {
+						return itemConverter.toBoundary(entity);
+					}
+				}) // Stream<ItemBoundary>
+				.collect(Collectors.toList()); // List<ItemBoundary>
 	}
 
 	@Override
-	public ItemBoundary[] getItemParents(String userSpace, String userEmail, String itemSpace, String itemId) {
-		//TODO
-		return null;
+	@Transactional(readOnly = true)
+	public List<ItemBoundary> getParents(String userSpace, String userEmail, String itemSpace, String itemId) {
+		ItemEntity childItem = this.itemDao.findById(new ItemId(itemSpace, itemId).toString())
+				.orElseThrow(() -> new ItemNotFoundException(
+						"item with id: " + itemId + "and space: " + itemSpace + " does not exist"));
+		Set<ItemEntity> itemParents = childItem.getItemParents();
+		return itemParents.stream() // Stream<ItemEntity>
+				.map(new Function<ItemEntity, ItemBoundary>() {
+					@Override
+					public ItemBoundary apply(ItemEntity entity) {
+						return itemConverter.toBoundary(entity);
+					}
+				}) // Stream<ItemBoundary>
+				.collect(Collectors.toList()); // List<ItemBoundary>
 	}
-
+	
 }
